Adicionei classe de validacao para entidade filme

+++ AvaliaFilmesAPI.Business/Service/FilmeService.cs
@@ -1,30 +1,48 @@
 ﻿using AvaliaFilmesAPI.Business.Service.Interface;
 using AvaliaFilmesAPI.Data.Repositories.InterfaceRepository;
 using AvaliaFilmesAPI.Domain.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using FluentValidation;
+using System.ComponentModel.DataAnnotations;
 
 namespace AvaliaFilmesAPI.Business.Service
 {
     public class FilmeService : IFilmeService
     {
         private readonly IFilmeRepository _filmeRepository;
+        private readonly FilmeValidation _filmeValidator;
 
-        public FilmeService(IFilmeRepository filmeRepository)
+        public FilmeService(IFilmeRepository filmeRepository, FilmeValidation filmeValidation)
         {
             _filmeRepository = filmeRepository;
+            _filmeValidator = filmeValidation;
         }
 
         public async Task AdicionarAsync(Filme filme)
         {
+            var validationResult = await _filmeValidator.ValidateAsync(filme);
+
+            if (!validationResult.IsValid)
+            {
+                throw new FluentValidation.ValidationException(validationResult.Errors);
+            }
+
             await _filmeRepository.AddAsync(filme);
         }
 
         public async Task AtualizarAsync(Filme filme)
         {
+            var validationResult = await _filmeValidator.ValidateAsync(filme);
+
+            if (!validationResult.IsValid)
+            {
+                throw new FluentValidation.ValidationException(validationResult.Errors);
+            }
+            var filmeExistente = await _filmeRepository.GetByIdAsync(filme.Id);
+
+            if (filmeExistente == null)
+            {
+                throw new ApplicationException($"Filme com Id '{filme.Id}' não encontrado.");
+            }
             await _filmeRepository.UpdateAsync(filme);
         }
 
@@ -32,6 +50,10 @@ namespace AvaliaFilmesAPI.Business.Service
         {
             var filme = await _filmeRepository.GetByIdAsync(id);
 
+            if (filme == null)
+            {
+                throw new Exception($"Filme com Id '{id}' não encontrado.");
+            }
             return filme;
         }
 
@@ -42,6 +64,12 @@ namespace AvaliaFilmesAPI.Business.Service
 
         public async Task RemoverAsync(Guid id)
         {
+            var filmeExistente = await _filmeRepository.GetByIdAsync(id);
+
+            if (filmeExistente == null)
+            {
+                throw new ApplicationException($"Filme com Id '{id}' não encontrado.");
+            }
             await _filmeRepository.DeleteAsync(id);
         }
     }
